Write some documentation

+++ src/Controller/UserController.php
@@ -7,13 +7,17 @@ use App\Repository\ClientRepository;
 use App\Repository\UserRepository;
 use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
+use JMS\Serializer\SerializationContext;
 use JMS\Serializer\SerializerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Entity;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class UserController extends AbstractController
 {
@@ -23,7 +27,8 @@ class UserController extends AbstractController
     public function getAllUsers(UserRepository $userRepository, SerializerInterface $serializer): JsonResponse
     {
         $usersList = $userRepository->findAll();
-        $jsonUsersLists = $serializer->serialize($usersList, 'json');
+        $context = SerializationContext::create()->setGroups(["getUsers"]);
+        $jsonUsersLists = $serializer->serialize($usersList, 'json', $context);
 
         return new JsonResponse($jsonUsersLists, Response::HTTP_OK, [], true);
     }
@@ -32,9 +37,10 @@ class UserController extends AbstractController
      * @Route("/api/users/{id}", name="app_user_details", methods="GET")
      * @Entity("user", expr="repository.find(id)")
      */
-    public function getUserDetails(User $user, SerializerInterface $serializer): JsonResponse
+    public function getUserDetails(User $user, SerializerInterface $serializer, ValidatorInterface $validator): JsonResponse
     {
-        $jsonUser = $serializer->serialize($user, 'json');
+        $context = SerializationContext::create()->setGroups(["getUsers"]);
+        $jsonUser = $serializer->serialize($user, 'json', $context);
 
         return new JsonResponse($jsonUser, Response::HTTP_OK, [], true);
     }
@@ -42,23 +48,24 @@ class UserController extends AbstractController
     /**
      * @Route("/api/users", name="app_create_user", methods="POST")
      */
-    public function createrUser(Request $request, SerializerInterface $serializer, ClientRepository $clientRepository, EntityManagerInterface $em): JsonResponse
+    public function createrUser(Request $request, SerializerInterface $serializer, ClientRepository $clientRepository, EntityManagerInterface $em, ValidatorInterface $validator): JsonResponse
     {
-        $clientList = [];
-
-        foreach ($clientRepository->findAll() as $client) {
-            $clientList[] = $client;
-        }
-
         $user = $serializer->deserialize($request->getContent(), User::class, 'json');
         $content = $request->toArray();
         $user->setEmail($content['email']);
         $user->setCreationDate(new DateTime());
-        $user->setClient($clientList[array_rand($clientList)]);
-        $jsonUser = $serializer->serialize($user, 'json');
+        $user->setClient($this->getUser());
+
+        $errors = $validator->validate($user);
+        if ($errors->count() > 0) {
+            return new JsonResponse($serializer->serialize($errors, 'json'), JsonResponse::HTTP_BAD_REQUEST, [], true);
+        }
 
         $em->persist($user);
         $em->flush();
+
+        $context = SerializationContext::create()->setGroups(["getUsers"]);
+        $jsonUser = $serializer->serialize($user, 'json', $context);
 
         return new JsonResponse($jsonUser, Response::HTTP_CREATED, [], true);
     }
@@ -69,6 +76,11 @@ class UserController extends AbstractController
      */
     public function deleteUser(User $user, EntityManagerInterface $em): JsonResponse
     {
+
+        if ($user->getClient() !== $this->getUser()) {
+            throw new AccessDeniedHttpException("Vous n'avez pas les droits pour supprimer cet utilisateur.");
+        }
+
         $em->remove($user);
         $em->flush();
